mgr/cephadm: fix tuned profiles getting removed if name has dashes
Previously, due to the way it was just splitting the file name on
dashes to try and get the profile name, any profile name with dashes
was not getting properly matched and would therefore get marked
stray and removed. This new strategy instead tries to match the actual
expected file name.

Fixes: https://tracker.ceph.com/issues/57405

Signed-off-by: Adam King <adking@redhat.com>